# openapi.yaml
openapi: 3.0.0
info:
  title: 'Laravel API'
  version: '1.0'
servers:
  -
    url: 'http://localhost:80'
paths:
  /api/login:
    post:
      tags:
        - Auth
      operationId: 99539a8e18179644cbac3165cf994fe5
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/health:
    get:
      tags:
        - health
      summary: 'Check the health of the API'
      operationId: 597094aa531969ea5921dd7d260a15e3
      responses:
        '200':
          description: OK
  /api/todos:
    get:
      tags:
        - Todo
      operationId: 403cd30b0c347e1f4f3cb2550035f43d
      parameters:
        -
          name: userID
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      tags:
        - Todo
      operationId: 9cf34bbb55f8bad045cdb2e4f8451890
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
  /api/users:
    post:
      tags:
        - User
      operationId: 5b6e72f178a035f354aff3d8b9eeb221
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
components:
  schemas:
    LoginInput:
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
      type: object
    CreateTodoInput:
      required:
        - userID
        - content
      properties:
        userID:
          type: integer
        content:
          type: string
      type: object
    Todo:
      required:
        - todoID
        - userID
        - content
        - createdAt
      properties:
        todoID:
          type: integer
        userID:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      type: object
    CreateUserInput:
      required:
        - email
        - name
        - password
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
      type: object
    User:
      required:
        - UserID
        - email
        - name
      properties:
        UserID:
          type: integer
        email:
          type: string
        name:
          type: string
      type: object
tags:
  -
    name: Auth
    description: Auth
  -
    name: health
    description: health
  -
    name: Todo
    description: Todo
  -
    name: User
    description: User